+++ examples/chapi-example/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @learncard/chapi-example
 
+## 1.0.233
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/init@2.0.18
+    -   @learncard/react@2.7.91
+
+## 1.0.232
+
+### Patch Changes
+
+-   Updated dependencies [[`d1c55ca4e179162ba0d0b63e85b0e16eebae3278`](https://github.com/learningeconomy/LearnCard/commit/d1c55ca4e179162ba0d0b63e85b0e16eebae3278)]:
+    -   @learncard/react@2.7.90
+
 ## 1.0.231
 
 ### Patch Changes