+++ services/learn-card-network/brain-service/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @learncard/network-brain-service
 
+## 3.4.1
+
+### Patch Changes
+
+-   [#690](https://github.com/learningeconomy/LearnCard/pull/690) [`933d49f9dacc2f376042bb1ed3843c9858620d3c`](https://github.com/learningeconomy/LearnCard/commit/933d49f9dacc2f376042bb1ed3843c9858620d3c) Thanks [@TaylorBeeston](https://github.com/TaylorBeeston)! - Do not allow receiving a credential or presentation more than once
+
 ## 3.4.0
 
 ### Minor Changes